+++ src/test/kotlin/com/zepben/evolve/services/network/tracing/phases/RemovePhasesTest.kt
@@ -11,13 +11,13 @@ package com.zepben.evolve.services.network.tracing.phases
 import com.zepben.evolve.cim.iec61970.base.core.ConductingEquipment
 import com.zepben.evolve.cim.iec61970.base.core.PhaseCode
 import com.zepben.evolve.cim.iec61970.base.core.Terminal
+import com.zepben.evolve.cim.iec61970.base.wires.EnergySource
 import com.zepben.evolve.services.network.NetworkService
 import com.zepben.evolve.services.network.tracing.networktrace.operators.NetworkStateOperators
 import com.zepben.evolve.services.network.tracing.phases.PhaseValidator.validatePhases
 import com.zepben.evolve.testing.TestNetworkBuilder
 import com.zepben.testutils.junit.SystemLogExtension
 import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.RegisterExtension
 import com.zepben.evolve.cim.iec61970.base.wires.SinglePhaseKind as SPK
@@ -36,7 +36,7 @@ internal class RemovePhasesTest {
     //
     // s4 --c5--
     //
-    private val n = TestNetworkBuilder()
+    private var n = TestNetworkBuilder()
         .fromSource(PhaseCode.ABCN) // s0
         .toAcls(PhaseCode.ABCN) // c1
         .toAcls(PhaseCode.ABCN) // c2
@@ -46,20 +46,20 @@ internal class RemovePhasesTest {
         .toAcls(PhaseCode.ABCN) // c5
         .build()
 
-    @BeforeEach
-    internal fun beforeEach() {
-        PhaseLogger.trace(listOf(n["s0"], n["s4"]))
+    @AfterEach
+    internal fun afterEach() {
+        PhaseLogger.trace(n.listOf<EnergySource>())
+    }
+
+    @Test
+    internal fun `validate base network`() {
+        // This is probably a pointless test, but it used to be rolled into a BeforeEach block and is now only executed once.
         validatePhases(n, "s0", PhaseCode.ABCN)
         validatePhases(n, "c1", PhaseCode.ABCN, PhaseCode.ABCN)
         validatePhases(n, "c2", PhaseCode.ABCN, PhaseCode.ABCN)
         validatePhases(n, "c3", PhaseCode.AB, PhaseCode.AB)
         validatePhases(n, "s4", PhaseCode.ABCN)
         validatePhases(n, "c5", PhaseCode.ABCN, PhaseCode.ABCN)
-    }
-
-    @AfterEach
-    internal fun afterEach() {
-        PhaseLogger.trace(listOf(n["s0"], n["s4"]))
     }
 
     @Test
@@ -102,6 +102,69 @@ internal class RemovePhasesTest {
         validatePhases(n.getT("s4", 1), PhaseCode.ABCN, PhaseCode.NONE)
         validatePhases(n.getT("c5", 1), PhaseCode.ABCN, PhaseCode.NONE)
         validatePhases(n.getT("c5", 2), PhaseCode.ABCN, PhaseCode.NONE)
+    }
+
+    @Test
+    internal fun `stops at open points`() {
+        //
+        // s0 11--c1--21 b2 21--c3--21 s4
+        //
+        n = TestNetworkBuilder()
+            .fromSource() // s0
+            .toAcls() // c1
+            .toBreaker(isNormallyOpen = true) // b2
+            .toAcls() // c3
+            .toSource() // s4
+            .build()
+
+        RemovePhases().run(n["s0-t1"]!!, NetworkStateOperators.NORMAL)
+        RemovePhases().run(n["s0-t1"]!!, NetworkStateOperators.CURRENT)
+
+        validatePhases(n, "s0", PhaseCode.NONE)
+        validatePhases(n, "c1", PhaseCode.NONE, PhaseCode.NONE)
+        validatePhases(n, "b2", PhaseCode.NONE, PhaseCode.ABC)
+        validatePhases(n, "c3", PhaseCode.ABC, PhaseCode.ABC)
+        validatePhases(n, "s4", PhaseCode.ABC)
+    }
+
+    @Test
+    internal fun `revisits equipment when ebbing different phases`() {
+        // This test was added when we stopped using a branching trace to ensure it still revisits items when removing different phases.
+        //
+        //               1------c7-----2
+        //               3             3
+        // s0 11--c1--21 j2 41--c3--21 j4 41--c5--
+        //               2             2
+        //               1------c6-----2
+        //
+        // If c5 isn't revisited it will have phases still as c3, c6 and c7 are only single phase lines,
+        //
+        n = TestNetworkBuilder()
+            .fromSource() // s0
+            .toAcls() // c1
+            .toJunction(numTerminals = 4) // j2
+            .toAcls(nominalPhases = PhaseCode.A) // c3
+            .toJunction(numTerminals = 4) // j4
+            .toAcls() // c5
+            .fromAcls(nominalPhases = PhaseCode.B) // c6
+            .connect("j2", "c6", 2, 1)
+            .connect("j4", "c6", 2, 2)
+            .fromAcls(nominalPhases = PhaseCode.C) // c7
+            .connect("j2", "c6", 2, 1)
+            .connect("j4", "c6", 2, 2)
+            .build()
+
+        RemovePhases().run(n["s0-t1"]!!, NetworkStateOperators.NORMAL)
+        RemovePhases().run(n["s0-t1"]!!, NetworkStateOperators.CURRENT)
+
+        validatePhases(n, "s0", PhaseCode.NONE)
+        validatePhases(n, "c1", PhaseCode.NONE, PhaseCode.NONE)
+        validatePhases(n, "j2", PhaseCode.NONE, PhaseCode.NONE, PhaseCode.NONE, PhaseCode.NONE)
+        validatePhases(n, "c3", PhaseCode.NONE, PhaseCode.NONE)
+        validatePhases(n, "j4", PhaseCode.NONE, PhaseCode.NONE, PhaseCode.NONE, PhaseCode.NONE)
+        validatePhases(n, "c5", PhaseCode.NONE, PhaseCode.NONE)
+        validatePhases(n, "c6", PhaseCode.NONE, PhaseCode.NONE)
+        validatePhases(n, "c7", PhaseCode.NONE, PhaseCode.NONE)
     }
 
     private fun NetworkService.getT(ce: String, t: Int): Terminal =